/internal/storage/domain: consolidate
